Disabling the Bootstrap fixed navbar
I can't immediately figure out the best way to make this work with the
Drupal tools overlay that also wants to float at the top of the page.
Maybe that's better address with an admin theme. For now this will make
the site work better in the non-logged in state.